PetHome.Domain находился в корневвой папке. Создал новую библетеку классов PetHome.Domain в src и перенес в неё всё содержимое
