--- Domain/GeneralValueObjects/FullName.cs
@@ -1,26 +1,0 @@
-﻿using CSharpFunctionalExtensions;
-using PetHome.Domain.Shared.Error;
-
-namespace PetHome.Domain.GeneralValueObjects;
-
-public record FullName
-{
-    public const int MAX_NAME_LENGTH = 20;
-    public string FirstName { get; }
-    public string LastName { get; }
-
-    private FullName() { }
-    private FullName(string firstName, string lastName)
-    {
-        FirstName = firstName;
-        LastName = lastName;
-    }
-
-    public static Result<FullName, Error> Create(string firstName, string lastName)
-    {
-        if (string.IsNullOrWhiteSpace(firstName) || string.IsNullOrWhiteSpace(lastName))
-            return Errors.Validation("Имя и фамилия");
-
-        return new FullName(firstName, lastName);
-    }
-}
