--- Domain/GeneralValueObjects/PaymentMethodEnum.cs
@@ -1,7 +1,0 @@
-﻿namespace PetHome.Domain.GeneralValueObjects;
-public enum PaymentMethodEnum
-{
-    Card,
-    Cash,
-    Help
-}
