--- Domain/GeneralValueObjects/SocialNetwork.cs
@@ -1,23 +1,0 @@
-﻿using CSharpFunctionalExtensions;
-using PetHome.Domain.Shared.Error;
-
-namespace PetHome.Domain.GeneralValueObjects;
-public record SocialNetwork
-{
-    public string Value { get; }
-
-    private SocialNetwork() { }
-    private SocialNetwork(string value)
-    {
-        Value = value;
-    }
-    public string Url { get; private set; }
-
-    public static Result<SocialNetwork, Error> Create(string value)
-    {
-        if (string.IsNullOrWhiteSpace(value))
-            return Errors.Validation("Социальная сеть");
-
-        return new SocialNetwork(value);
-    }
-}
