--- Domain/PetEntity/BreedName.cs
@@ -1,22 +1,0 @@
-﻿using CSharpFunctionalExtensions;
-using PetHome.Domain.Shared.Error;
-
-namespace PetHome.Domain.PetEntity;
-public record BreedName
-{ 
-    public string Value { get; }
-
-    private BreedName() { }
-    private BreedName(string value)
-    {
-        Value = value;
-    }
-
-    public static Result<BreedName, Error> Create(string value)
-    {
-        if (string.IsNullOrWhiteSpace(value))
-            return Errors.Validation("Порода");
-
-        return new BreedName(value);
-    }
-}
