--- Domain/PetEntity/Color.cs
@@ -1,27 +1,0 @@
-﻿using CSharpFunctionalExtensions;
-using PetHome.Domain.Shared.Error;
-
-namespace PetHome.Domain.PetEntity;
-public class Color : ValueObject
-{
-    public string Value { get; }
-
-    private Color() { }
-    private Color(string value)
-    {
-        Value = value;
-    }
-
-    public static Result<Color, Error> Create(string value)
-    {
-        if (string.IsNullOrWhiteSpace(value))
-            return Errors.Validation("Цвет");
-
-        return new Color(value);
-    }
-
-    protected override IEnumerable<object> GetEqualityComponents()
-    {
-        yield return Value;
-    }
-}
