--- Domain/PetEntity/PetStatusEnum.cs
@@ -1,7 +1,0 @@
-﻿namespace PetHome.Domain.PetEntity;
-public enum PetStatusEnum
-{
-    isFree,
-    isTreatment,
-    isHomed,
-}
