--- Domain/Shared/Constants.cs
@@ -1,7 +1,0 @@
-﻿namespace PetHome.Domain.Shared;
-public static class Constants
-{
-    public static int MAX_NAME_LENGHT = 20;
-    public static int MAX_DESC_LENGHT = 1000;
-    public static int MAX_WEIGHT = 500;
-}
