+++ PetHome.sln
@@ -9,9 +9,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PetHome.Infrastructure", "s
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PetHome.Application", "src\PetHome.Application\PetHome.Application.csproj", "{0048BDF5-B335-4749-8D8C-AC319E79C9BF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PetHome.Domain", "Domain\PetHome.Domain.csproj", "{ED63A11F-D618-4F35-875F-D75B293D03BF}"
-EndProject
 Project("{E53339B2-1760-4266-BCC7-CA923CBCF16C}") = "docker-compose", "docker-compose.dcproj", "{93FD1EE1-20EB-42CE-BCB1-0B3BEC26F4AB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PetHome.Domain", "src\PetHome.Domain\PetHome.Domain.csproj", "{D653C3FA-AA77-4E89-839A-B0BA44A785B5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -31,14 +31,14 @@ Global
 		{0048BDF5-B335-4749-8D8C-AC319E79C9BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0048BDF5-B335-4749-8D8C-AC319E79C9BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0048BDF5-B335-4749-8D8C-AC319E79C9BF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{ED63A11F-D618-4F35-875F-D75B293D03BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ED63A11F-D618-4F35-875F-D75B293D03BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ED63A11F-D618-4F35-875F-D75B293D03BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{ED63A11F-D618-4F35-875F-D75B293D03BF}.Release|Any CPU.Build.0 = Release|Any CPU
 		{93FD1EE1-20EB-42CE-BCB1-0B3BEC26F4AB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{93FD1EE1-20EB-42CE-BCB1-0B3BEC26F4AB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{93FD1EE1-20EB-42CE-BCB1-0B3BEC26F4AB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{93FD1EE1-20EB-42CE-BCB1-0B3BEC26F4AB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D653C3FA-AA77-4E89-839A-B0BA44A785B5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D653C3FA-AA77-4E89-839A-B0BA44A785B5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D653C3FA-AA77-4E89-839A-B0BA44A785B5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D653C3FA-AA77-4E89-839A-B0BA44A785B5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
