+++ src/PetHome.API/PetHome.API.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
@@ -27,8 +27,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Domain\PetHome.Domain.csproj" />
     <ProjectReference Include="..\PetHome.Application\PetHome.Application.csproj" />
+    <ProjectReference Include="..\PetHome.Domain\PetHome.Domain.csproj" />
     <ProjectReference Include="..\PetHome.Infrastructure\PetHome.Infrastructure.csproj" />
   </ItemGroup>
