+++ src/PetHome.Domain/GeneralValueObjects/Email.cs
