+++ src/PetHome.Domain/GeneralValueObjects/FullName.cs
@@ -1,20 +1,26 @@
-﻿using PetHome.Domain.VolunteerEntity;
+﻿using CSharpFunctionalExtensions;
+using PetHome.Domain.Shared.Error;
 
-namespace PetHome.Domain.GeneralValueObjects
+namespace PetHome.Domain.GeneralValueObjects;
+
+public record FullName
 {
-    public record FullName
+    public const int MAX_NAME_LENGTH = 20;
+    public string FirstName { get; }
+    public string LastName { get; }
+
+    private FullName() { }
+    private FullName(string firstName, string lastName)
     {
-        public const int MAX_NAME_LENGTH = 20;
+        FirstName = firstName;
+        LastName = lastName;
+    }
 
-        public string FirstName { get; }
-        public string LastName { get; }
-        private FullName(string firstName, string lastName)
-        {
-            FirstName = firstName;
-            LastName = lastName;
-        }
+    public static Result<FullName, Error> Create(string firstName, string lastName)
+    {
+        if (string.IsNullOrWhiteSpace(firstName) || string.IsNullOrWhiteSpace(lastName))
+            return Errors.Validation("Имя и фамилия");
 
-        public static FullName Create(string firstName, string lastName) => new FullName(firstName, lastName);
-
+        return new FullName(firstName, lastName);
     }
 }
