+++ src/PetHome.Domain/GeneralValueObjects/PhoneNumber.cs
@@ -1,30 +1,24 @@
 ﻿using CSharpFunctionalExtensions;
+using PetHome.Domain.Shared.Error;
 using System.Text.RegularExpressions;
 
 namespace PetHome.Domain.GeneralValueObjects;
-public class PhoneNumber : ValueObject
-{
+public record PhoneNumber
+{ 
     private const string PhoneNumberRegex = @"^\(?([0-9]{3})\)?[-.●]?([0-9]{3})[-.●]?([0-9]{4})$";
     public string Value { get; }
 
+    private PhoneNumber() { }
     private PhoneNumber(string value)
     {
         Value = value;
     }
 
-    public static Result<PhoneNumber> Create(string value)
+    public static Result<PhoneNumber, Error> Create(string value)
     {
-        if (string.IsNullOrWhiteSpace(value))
-            return Result.Failure<PhoneNumber>("Введите номер телефона");
-
-        if (!Regex.IsMatch(value, PhoneNumberRegex))
-            return Result.Failure<PhoneNumber>("Номер телефона не соответствует формату");
+        if (string.IsNullOrWhiteSpace(value))//|| (!Regex.IsMatch(value, PhoneNumberRegex)))
+            return Errors.Validation("Номер телефона");
 
         return new PhoneNumber(value);
-    }
-
-    protected override IEnumerable<object> GetEqualityComponents()
-    {
-        yield return Value;
     }
 }
