+++ src/PetHome.Domain/GeneralValueObjects/PhoneNumbersDetails.cs
