+++ src/PetHome.Domain/GeneralValueObjects/Requisites.cs
@@ -1,12 +1,14 @@
 ﻿using CSharpFunctionalExtensions;
+using PetHome.Domain.Shared.Error;
 
 namespace PetHome.Domain.GeneralValueObjects;
-public class Requisites : ValueObject
+public record Requisites
 {
     public string Name { get; }
     public string Description { get; }
     public PaymentMethodEnum PaymentMethod { get; }
 
+    private Requisites() { }
     private Requisites(string name, string description, PaymentMethodEnum paymentMethod)
     {
         Name = name;
@@ -14,19 +16,13 @@ public class Requisites : ValueObject
         PaymentMethod = paymentMethod;
     }
 
-    public static Result<Requisites> Create(string name, string description, PaymentMethodEnum paymentMethod)
+    public static Result<Requisites, Error> Create(string name, string description, PaymentMethodEnum paymentMethod)
     {
         bool isInvalidRequisite = string.IsNullOrEmpty(name) || string.IsNullOrEmpty(description) || paymentMethod == null;
 
         if (isInvalidRequisite)
-            return Result.Failure<Requisites>("Некорректно введены реквезиты");
+            return Errors.Validation("Реквезиты");
 
         return new Requisites(name, description, paymentMethod);
-    }
-
-    protected override IEnumerable<object> GetEqualityComponents()
-    {
-        yield return Name;
-        yield return PaymentMethod;
     }
 }
