+++ src/PetHome.Domain/GeneralValueObjects/RequisitesDetails.cs
