+++ src/PetHome.Domain/GeneralValueObjects/SocialNetworkDetails.cs
