--- src/PetHome.Domain/GeneralValueObjects/VO_Date.cs
@@ -1,27 +1,0 @@
-﻿using CSharpFunctionalExtensions;
-
-namespace PetHome.Domain.GeneralValueObjects;
-public class VO_Date : ValueObject
-{
-    public DateOnly Value { get; }
-
-    private VO_Date(DateOnly value)
-    {
-        Value = value;
-    }
-
-    public static Result<VO_Date> Create(DateOnly value)
-    {
-        if (value == null || value.Year > 100)
-        {
-            return Result.Failure<VO_Date>("Введите корректную дату");
-        }
-
-        return new VO_Date(value);
-    }
-
-    protected override IEnumerable<object> GetEqualityComponents()
-    {
-        yield return Value;
-    }
-}
