+++ src/PetHome.Domain/PetEntity/Breed.cs
@@ -2,12 +2,10 @@
 
 namespace PetHome.Domain.PetEntity;
 public class Breed
-{
-    // Для EF core
-    public Breed() { }
+{ 
+    private Breed() { }
 
     public BreedId Id { get; private set; }
     public BreedName Name { get; private set; }
-    public SpeciesId SpeciesId { get; private set; }
-
+    public SpeciesId SpeciesId { get; private set; } 
 }
