+++ src/PetHome.Domain/PetEntity/PetId.cs
@@ -1,27 +1,25 @@
-﻿using CSharpFunctionalExtensions;
-
-namespace PetHome.Domain.PetEntity
-{
-    public record class PetId
+﻿namespace PetHome.Domain.PetEntity;
+public record class PetId
+{  
+    public Guid Value { get; }
+    
+    private PetId() { }
+    private PetId(Guid value)
     {
-        public Guid Value { get; }
-        private PetId(Guid value)
-        {
-            Value = value;
-        }
+        Value = value;
+    }
 
-        public static PetId Create() => new PetId(Guid.NewGuid()); 
-                
-        public static PetId Create(Guid id) => new PetId(id); 
-                
-        public static PetId CreateEmpty() => new PetId(Guid.Empty); 
+    public static PetId Create() => new PetId(Guid.NewGuid()); 
+            
+    public static PetId Create(Guid id) => new PetId(id); 
+            
+    public static PetId CreateEmpty() => new PetId(Guid.Empty); 
 
-        public static implicit operator Guid(PetId petId)
-        {
-            if (petId == null)
-                throw new ArgumentNullException();
+    public static implicit operator Guid(PetId petId)
+    {
+        if (petId == null)
+            throw new ArgumentNullException();
 
-            return petId.Value;
-        }
+        return petId.Value;
     }
 }
