+++ src/PetHome.Domain/PetEntity/PetShelter.cs
@@ -1,8 +1,7 @@
 ﻿namespace PetHome.Domain.PetEntity;
 public class PetShelter
-{
-    // Для EF core
-    public PetShelter() { }
+{ 
+    private PetShelter() { }
 
     public PetShelterId Id { get; private set; }
     public ShelterName Name { get; private set; }
