+++ src/PetHome.Domain/PetEntity/PetShelterId.cs
@@ -1,27 +1,25 @@
-﻿using CSharpFunctionalExtensions;
-
-namespace PetHome.Domain.PetEntity
+﻿namespace PetHome.Domain.PetEntity;
+public record PetShelterId
 {
-    public record PetShelterId
+    public Guid Value { get; }
+
+    private PetShelterId() { } 
+    private PetShelterId(Guid value)
     {
-        public Guid Value { get; }
-        private PetShelterId(Guid value)
-        {
-            Value = value;
-        }
+        Value = value;
+    }
 
-        public static PetShelterId Create() => new PetShelterId(Guid.NewGuid()); 
+    public static PetShelterId Create() => new PetShelterId(Guid.NewGuid()); 
 
-        public static  PetShelterId Create(Guid id) => new PetShelterId(id); 
+    public static  PetShelterId Create(Guid id) => new PetShelterId(id); 
 
-        public static PetShelterId CreateEmpty() => new PetShelterId(Guid.Empty); 
+    public static PetShelterId CreateEmpty() => new PetShelterId(Guid.Empty); 
 
-        public static implicit operator Guid(PetShelterId petShelterId)
-        {
-            if (petShelterId == null)
-                throw new ArgumentNullException();
+    public static implicit operator Guid(PetShelterId petShelterId)
+    {
+        if (petShelterId == null)
+            throw new ArgumentNullException();
 
-            return petShelterId.Value;
-        }
+        return petShelterId.Value;
     }
 }
