+++ src/PetHome.Domain/PetEntity/Species.cs
@@ -1,13 +1,9 @@
-﻿using CSharpFunctionalExtensions;
-
-namespace PetHome.Domain.PetEntity;
-public record Species
-{ 
-    // Для EF core
+﻿namespace PetHome.Domain.PetEntity;
+public class Species
+{  
     public Species() { }
 
     public SpeciesId Id { get; private set; }
     public SpeciesName Name { get; private set; }
-    public IReadOnlyList<Breed> BreedList { get; private set; }
-
+    public IReadOnlyList<Breed> Breeds { get; private set; } 
 }
