+++ src/PetHome.Domain/Shared/Constants.cs
@@ -1,15 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace PetHome.Domain.Shared
+﻿namespace PetHome.Domain.Shared;
+public static class Constants
 {
-    public static class Constants
-    {
-        public static int MAX_NAME_LENGHT = 20;
-        public static int MAX_DESC_LENGHT = 1000;
-        public static int MAX_WEIGHT = 500;
-    }
+    public static int MAX_NAME_LENGHT = 20;
+    public static int MAX_DESC_LENGHT = 1000;
+    public static int MAX_WEIGHT = 500;
 }
