+++ src/PetHome.Domain/Shared/Error/Error.cs
