+++ src/PetHome.Domain/Shared/Error/ErrorType.cs
