+++ src/PetHome.Domain/Shared/Error/Errors.cs
