+++ src/PetHome.Domain/VolunteerEntity/Volunteer.cs
@@ -1,41 +1,67 @@
 ﻿using CSharpFunctionalExtensions;
 using PetHome.Domain.GeneralValueObjects;
 using PetHome.Domain.PetEntity;
+using PetHome.Domain.Shared.Error;
 
 namespace PetHome.Domain.VolunteerEntity;
 public class Volunteer
 {
-    //Для EF core
-    public Volunteer() { }
+    private Volunteer() { }
 
-    private Volunteer(FullName fullName, string email, string description, DateOnly startVolunteeringDate, 
-        PhoneNumber phoneNumber,/* List<SocialNetwork> socialNetworkList,*/ Requisites requisites) { }
+    private Volunteer( 
+        VolunteerId id,
+        FullName fullName,
+        Email email,
+        string description,
+        Date startVolunteeringDate,
+        PhoneNumbersDetails phoneNumbersDetails,
+        SocialNetworkDetails socialNetworkDetails,
+        RequisitesDetails requisitesDetails)
+    {
+        Id = id;
+        FullName = fullName;
+        Email = email;
+        Description = description;
+        StartVolunteeringDate = startVolunteeringDate;
+        PhoneNumberDetails = phoneNumbersDetails;
+        SocialNetworkDetails = socialNetworkDetails;
+        RequisitesDetails = requisitesDetails;
+
+    }
 
     public VolunteerId Id { get; private set; }
     public FullName FullName { get; private set; }
-    public string? Email { get; private set; }
+    public Email? Email { get; private set; }
     public string Description { get; private set; }
-    public DateOnly StartVolunteeringDate { get; private set; }
-    public IReadOnlyList<Pet> PetList { get; private set; }
+    public Date StartVolunteeringDate { get; private set; }
+    public IReadOnlyList<Pet> Pets { get; private set; }
     public int HomedPetsCount => GetPetCountByStatusAndVolunteer(PetStatusEnum.isHomed);
     public int FreePetsCount => GetPetCountByStatusAndVolunteer(PetStatusEnum.isFree);
     public int TreatmentPetsCount => GetPetCountByStatusAndVolunteer(PetStatusEnum.isTreatment);
-    public PhoneNumber PhoneNumber { get; private set; }
-   // public IReadOnlyList<SocialNetwork> SocialNetworkList { get; private set; }
-    public Requisites Requisites { get; private set; }
-    // TODO: Реализовать метод после конфигурации БД
-    private int GetPetCountByStatusAndVolunteer(PetStatusEnum status) => PetList.Where(pet => pet.Status == status && pet.VolunteerId == Id).Count();
+    public PhoneNumbersDetails? PhoneNumberDetails { get; private set; }
+    public RequisitesDetails? RequisitesDetails { get; private set; }
+    public SocialNetworkDetails? SocialNetworkDetails { get; private set; }
 
-    public Result<Volunteer> Create(FullName fullName, string email, string description, DateOnly startVolunteeringDate,
-        PhoneNumber phoneNumber, /*List<SocialNetwork> socialNetworkList, */Requisites requisites)
+
+    private int GetPetCountByStatusAndVolunteer(PetStatusEnum status) => Pets.Where(pet => pet.Status == status && pet.VolunteerId == Id).Count();
+
+    public static Result<Volunteer, Error> Create(
+        VolunteerId id,
+        FullName fullName,
+        Email email,
+        string description,
+        Date startVolunteeringDate,
+        PhoneNumbersDetails phoneNumbersDetails,
+        SocialNetworkDetails socialNetworkDetails,
+        RequisitesDetails requisitesDetails)
     {
         if (string.IsNullOrWhiteSpace(description))
-            return Result.Failure<Volunteer>("Введите описание");
+            return Errors.Validation("Описание");
 
         if (startVolunteeringDate == null)
-            return Result.Failure<Volunteer>("Выберите дату начала волонтёрства");
+            return Errors.Validation("Дата начала волонтёрства");
 
 
-        return new Volunteer(fullName, email, description, startVolunteeringDate, phoneNumber,/* socialNetworkList,*/ requisites) { };
+        return new Volunteer(id, fullName, email, description, startVolunteeringDate, phoneNumbersDetails, socialNetworkDetails, requisitesDetails) { };
     }
 }
