+++ src/PetHome.Infrastructure/PetHome.Infrastructure.csproj
@@ -23,8 +23,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Domain\PetHome.Domain.csproj" />
     <ProjectReference Include="..\PetHome.Application\PetHome.Application.csproj" />
+    <ProjectReference Include="..\PetHome.Domain\PetHome.Domain.csproj" />
   </ItemGroup>
 
 </Project>
